Created page folder and moved home page
